+++ federation/protocols/diaspora/encrypted.py
@@ -1,7 +1,8 @@
 import json
-from base64 import b64decode
+from base64 import b64decode, b64encode
 
 from Crypto.Cipher import PKCS1_v1_5, AES
+from Crypto.Random import get_random_bytes
 from lxml import etree
 
 
@@ -28,3 +29,33 @@ class EncryptedPayload:
         encrypter = AES.new(key, AES.MODE_CBC, iv)
         content = encrypter.decrypt(encrypted_magic_envelope)
         return etree.fromstring(pkcs7_unpad(content))
+
+    @staticmethod
+    def get_aes_key_json():
+        iv = get_random_bytes(AES.block_size)
+        key = get_random_bytes(32)
+        encrypter = AES.new(key, AES.MODE_CBC, iv)
+        return {
+            "key": b64encode(key),
+            "iv": b64encode(iv),
+        }, encrypter
+
+    @staticmethod
+    def encrypt(payload, public_key):
+        """
+        Encrypt a payload using an encrypted JSON wrapper.
+
+        See: <insert link to docs>
+
+        :param payload: Payload document as a string.
+        :param public_key: Public key of recipient as an RSA object.
+        :return: Encrypted JSON wrapper as dict.
+        """
+        aes_key_json, encrypter = EncryptedPayload.get_aes_key_json()
+        encrypted_me = encrypter.encrypt(payload)
+        cipher = PKCS1_v1_5.new(public_key)
+        aes_key = cipher.encrypt(aes_key_json)
+        return {
+            "aes_key": b64encode(aes_key),
+            "encrypted_magic_envelope": b64encode(encrypted_me),
+        }